feat: Implement entity movement and collision detection
This commit introduces the ability for entities to move around the map and interact with terrain.

- Implemented the `TryMove` command, which now checks for collisions with walls and updates the entity's position if the move is valid.
- Updated the `execute_actions` system to use the `TryMove` command for `Action::Move`.
- Added logging to `TryMove` to indicate when a move is attempted and when a wall is bumped into.
- The `Map` resource now stores the terrain type of each tile.
- The `spawn_player` system now spawns an enemy.
- The `keyboard_input` system now correctly handles the player's input and schedules the next turn.
- The `turn_queue` now correctly handles wrapping time.
- The `position` component now correctly handles adding a move direction.
- The `action` component now has an `InputAction` enum.

These changes lay the groundwork for more complex interactions and game mechanics.

+++ src/model/components/turn.rs
@@ -5,7 +5,7 @@ use super::MoveDirection;
 // Define components for the turn system
 #[derive(Component, Debug)]
 pub struct TurnActor {
-    pub speed: u32,
+    pub speed: u64,
     pub next_turn_time: u64,
 }
